Initial XML reader for header btns
In order to make the page easier to edit, the initial XML structure was created to make a dynamic header link bar. Hopefully this should make it so only the XML file needs to be changed rather than the entire site. The XML parser was copied from techdump.html, with some cosmetic changes. In the future the plan is to consolidate the XML reading functions so that the site is more organized.