Temp fix for link issues
Removed broken links in header of techdump.html, the header is still not cooperating with the other XML reading function used on this page, so the broken links have been removed and the functionality will be fixed later. Additionally, the orange button had the old link for the TDS, so it was updated.